+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/level/block/BlockFeatures.java
@@ -3,6 +3,10 @@ package net.potionstudios.biomeswevegone.world.level.block;
 import net.minecraft.world.level.ItemLike;
 import net.minecraft.world.level.block.*;
 import net.potionstudios.biomeswevegone.world.item.BWGItems;
+import net.potionstudios.biomeswevegone.world.level.block.plants.bush.DesertPlantBlock;
+import net.potionstudios.biomeswevegone.world.level.block.plants.cactus.BWGCactusBlock;
+import net.potionstudios.biomeswevegone.world.level.block.plants.vegetation.AloeVeraBlock;
+import net.potionstudios.biomeswevegone.world.level.block.plants.vegetation.BWGDoublePlantBlock;
 import net.potionstudios.biomeswevegone.world.level.block.wood.BWGWood;
 import net.potionstudios.biomeswevegone.world.level.block.wood.BWGWoodSet;
 import org.apache.logging.log4j.util.TriConsumer;
@@ -18,11 +22,11 @@ public class BlockFeatures {
     public static void registerCompostables(BiConsumer<ItemLike, Float> consumer) {
         BWGBlocks.BLOCKS.forEach(object -> {
             Block block = object.get();
-            if (block instanceof TallGrassBlock)
+            if (block instanceof TallGrassBlock || block instanceof BWGDoublePlantBlock)
                 consumer.accept(block, 0.3F);
-            else if (block instanceof FlowerBlock || block instanceof TallFlowerBlock || block instanceof WaterlilyBlock || block instanceof MushroomBlock)
+            else if (block instanceof FlowerBlock || block instanceof TallFlowerBlock || block instanceof WaterlilyBlock || block instanceof MushroomBlock || block instanceof AloeVeraBlock)
                 consumer.accept(block, 0.65F);
-            else if (block instanceof VineBlock)
+            else if (block instanceof VineBlock || block instanceof DesertPlantBlock || block instanceof BWGCactusBlock)
                 consumer.accept(block, 0.5F);
         });