+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/level/block/plants/cactus/OddityCactusBlock.java
@@ -1,7 +1,0 @@
-package net.potionstudios.biomeswevegone.world.level.block.plants.cactus;
-
-public class OddityCactusBlock extends BWGCactusBlock {
-	public OddityCactusBlock(Properties properties) {
-		super(properties);
-	}
-}