+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix Grass Generating in caves in Allium Shrubland
 - Adjust Just Like Grandma's Lang to say both pies
 - Make Stairs and Slabs Tags for BWG SandStone Sets and add to Respected Fabric Tags
+- Make BWG Desert Plants Compostable
 
 # 1.0.6-Beta
 - Fix Skyris Hanging Sign Textures being swapped